Move otelhttp to lib/observability/
